+++ src/Providers/ConditionalProvidersServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace SebastiaanLuca\ConditionalProviders\Providers;
 
+use Illuminate\Foundation\AliasLoader;
 use Illuminate\Support\ServiceProvider;
 
 class ConditionalProvidersServiceProvider extends ServiceProvider
@@ -13,14 +14,16 @@ class ConditionalProvidersServiceProvider extends ServiceProvider
     {
         $environment = $this->app->environment();
 
-        $providers = $this->app['config']->get('app.' . $environment . '_providers');
-
-        if (! $providers) {
-            return;
-        }
+        $providers = $this->app['config']->get('app.' . $environment . '_providers', []);
 
         foreach ($providers as $provider) {
             $this->app->register($provider);
+        }
+
+        $facades = $this->app['config']->get('app.' . $environment . '_aliases', []);
+
+        foreach ($facades as $class => $name) {
+            AliasLoader::getInstance()->alias($class, $name);
         }
     }
 }