+++ tests/TestCase.php
@@ -3,9 +3,6 @@
 namespace SebastiaanLuca\ConditionalProviders\Tests;
 
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
-use SebastiaanLuca\ConditionalProviders\Providers\ConditionalProvidersServiceProvider;
-use SebastiaanLuca\ConditionalProviders\Tests\Helpers\MyLocalServiceProvider;
-use SebastiaanLuca\ConditionalProviders\Tests\Helpers\MyStagingServiceProvider;
 
 class TestCase extends OrchestraTestCase
 {
@@ -17,9 +14,7 @@ class TestCase extends OrchestraTestCase
     /**
      * Resolve application core configuration implementation.
      *
-     * @param  \Illuminate\Foundation\Application $app
-     *
-     * @return void
+     * @param \Illuminate\Foundation\Application $app
      */
     protected function resolveApplicationConfiguration($app)
     {
@@ -28,17 +23,5 @@ class TestCase extends OrchestraTestCase
         $app->detectEnvironment(function () {
             return $this->environment;
         });
-
-        $app['config']->push('app.providers', ConditionalProvidersServiceProvider::class);
-
-        $app['config']->set('app.local_providers', [
-            MyLocalServiceProvider::class,
-        ]);
-
-        $app['config']->set('app.staging_providers', [
-            MyStagingServiceProvider::class,
-        ]);
-
-        $app['config']->set('app.testing_providers', []);
     }
 }