+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## Unreleased
 
+## 1.1.0 (2017-07-19)
+
+- Added conditional loading of aliases (facades)
+- Refactored and fixed tests
+
 ## 1.0.0 (2017-07-13)
 
 ### Added